Refactored project structure and code so each game is in its own folder
