+++ api/__init__.py
@@ -1,3 +1,3 @@
-from .wiki_api import WikiAPI, get_ongoing_events_async, get_wuwa_events_async, get_zzz_events_async
+from .wiki_api import WikiAPI, get_ongoing_events_async
 
-__all__ = ["WikiAPI", "get_ongoing_events_async", "get_wuwa_events_async", "get_zzz_events_async"]
+__all__ = ["WikiAPI", "get_ongoing_events_async"]
