+++ config/__init__.py
@@ -1,3 +1,3 @@
-from .config import TOKEN, GUILD_ID, GUILD_OBJECT, API_URL_WUWA, API_URL_ZZZ, GAME_CONFIG
+from .config import TOKEN, GUILD_ID, GUILD_OBJECT
 
-__all__ = ["TOKEN", "GUILD_ID", "GUILD_OBJECT", "API_URL_WUWA", "API_URL_ZZZ", "GAME_CONFIG"]
+__all__ = ["TOKEN", "GUILD_ID", "GUILD_OBJECT"]
