+++ games/wuwa/config.py
@@ -1,0 +1,11 @@
+import discord
+
+WUWA_CONFIG = {
+    "display_name": "Wuthering Waves",
+    "api_url": "https://wutheringwaves.fandom.com/api.php",
+    "category": "Events",
+    "color": discord.Color.blue(),
+    "emoji": "🌊",
+    "thumbnail_path": "images/WutheringWavesThumbnail.jpeg",
+    "thumbnail_filename": "WuWaThumbnail.png",
+}
