+++ games/zzz/config.py
@@ -1,0 +1,11 @@
+import discord
+
+ZZZ_CONFIG = {
+    "display_name": "Zenless Zone Zero",
+    "api_url": "https://zenless-zone-zero.fandom.com/api.php",
+    "category": "In-Game_Events",
+    "color": discord.Color.orange(),
+    "emoji": "⚡",
+    "thumbnail_path": "images/ZenlessZoneZeroThumbnail.png",
+    "thumbnail_filename": "ZZZThumbnail.png",
+}
